+++ compiler/reflaxe.php/test/compiler-tracer/expected/bootstrap.php
@@ -1,0 +1,6 @@
+<?php
+
+declare(strict_types=1);
+
+require_once __DIR__ . '/modules/6_tracer/4_Main/4_Main.php';
+Hx_6_tracer_4_Main::main();