+++ compiler/reflaxe.php/test/semantic-matrix/expected/modules/9_semantics/4_Main/4_Main.php
@@ -2,12 +2,6 @@
 
 declare(strict_types=1);
 
-class Hx_9_semantics_10_Calculator {
-	public static function add(int $left, int $right): int {
-		return $left + $right;
-	}
-}
-
 class Hx_9_semantics_4_Main {
 	public static function main(): void {
 		$answer = Hx_9_semantics_10_Calculator::add( 40, 2 );
@@ -40,5 +34,3 @@ class Hx_9_semantics_4_Main {
 		}
 	}
 }
-
-Hx_9_semantics_4_Main::main();